Move todo conversion out of list converters.
